+++ src/main/java/software/bernie/geckolib3/core/manager/AnimationData.java
@@ -15,10 +15,10 @@ public class AnimationData
 {
 	private HashMap<IBone, BoneSnapshot> boneSnapshotCollection;
 	private HashMap<String, AnimationController> animationControllers = new HashMap<>();
-	public float tick;
+	public double tick;
 	public boolean isFirstTick = true;
 	private double resetTickLength = 30;
-	public Float startTick;
+	public Double startTick;
 	public Object ticker;
 	public boolean shouldPlayWhilePaused = false;
